Reorganized some directories and added the start of Mongo support
