+++ package.json
@@ -18,6 +18,7 @@
     "cron-parser": "^4.9.0",
     "discord.js": "^14.14.1",
     "dotenv": "^16.3.1",
+    "mongodb": "^6.12.0",
     "node-cron": "^3.0.3",
     "nodemon": "^3.0.3",
     "typescript": "^5.3.3"
