+++ src/games/hk_starrail/checkin_hkstr.ts
@@ -1,4 +1,4 @@
-import {Profile} from "../../models";
+import {Profile} from "../../types";
 
 export async function hkstrCheckin(profile: Profile): Promise<string> {
