+++ tutorial_images/Hoyolab Daily checkin Bot instructions.pdf
